+++ util/mkReadme.js
@@ -2,4 +2,11 @@
 
 const fs = require('fs');
 
-fs.writeFileSync('README.md', `${fs.readFileSync('README.src.md')}\n\n${fs.readFileSync('CHANGELOG.md')}`);
+fs.writeFileSync('README.md', `${fs.readFileSync('README.src.md')}
+
+\`\`\`javascript
+${fs.readFileSync('example.js')}
+\`\`\`
+
+
+${fs.readFileSync('CHANGELOG.md')}`);